backend/core-service/cflow_backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
cryptography==44.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
docker==7.1.0
drf-polymorphic==2.1.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
msgpack==1.1.0
mysqlclient==2.2.7
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-decouple==3.8
python3-openid==3.2.0
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
smmap==5.0.2
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
urllib3==2.4.0
